moved tetris logic to core_game
